% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_agilent.R
\name{parse_agilent_spectrum_csv}
\alias{parse_agilent_spectrum_csv}
\title{Parse the oddly formatted .csv files exported by Agilent's MassHunter software into an R data frame.}
\usage{
parse_agilent_spectrum_csv(filename)
}
\arguments{
\item{filename}{A string with the path of the relevant .csv file.}
}
\value{
A data frame with columns "V2", "V3", "file" and "intensity"
}
\description{
Parse the oddly formatted .csv files exported by Agilent's MassHunter software into an R data frame.
}

